requirements.txt
streamlit
streamlit-javascript
Pillow
